Refact: Project overhaul
The project is now a collection of libraries, primarily a sandboxing
place. Design philosophy is still somewhat the code being a library
if later reuse could come in handy, but currently the project does not
support it out of the box. In other words, there is no packaging.

Feat:
- Added example codes to illustrate the usage of the dependencies
- Minimal working example for ChaiScript in a separate compilation unit

Refact:
- Moving library code to Nova
- `deps` folder for git submodules
  - Added project Yui
  - ChaiScript moved under `deps`
- Old dependencies cleaned-up
  - `spdlog` removed, will be used via packaged manager if needed
- Removed CMake install related files (will be reintroduced later with a
  clean design if needed)
- Removed Conan dependency
- Removed example project